Create PaintCalculator
Program is written in C#. Idea is to calculate gallons of paint needed to paint the ceiling of a room. I prompt for the length and width, and assume one gallon covers 350 square feet. I chose to display the number of gallons needed to paint the ceiling as a whole number. User can also choose to calculate needed paint for a round room.